core/state: cleanup
